refacto(reset-password): code enhancement last part
